requirements_py3.10.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.5.2
apiview==1.3.25
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
diff-match-patch==20241021
distro==1.9.0
Django==4.2.21
django-import-export==4.1.1
django-jsonfield-backport==1.0.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
flower==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.10.0
idna==3.10
jiter==0.9.1
kombu==5.5.4
openai==2.0.0
packaging==25.0
pillow==10.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.52
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyMySQL==1.1.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.5.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
